Split compare page helper modules
